widget: rename `widget_settings` -> `settings` mod
To stay consistent with the rest of the concise naming "conventions"
within the `widget` module.
